requirements.txt
langchain==0.3.26
langchain_core==0.3.66
langchain_deepseek==0.1.3
langchain_openai==0.3.27
memory==1.0.0
pandas==2.3.0
pydantic==2.11.7
python-dotenv==1.1.1
rich==14.0.0
termcolor==3.1.0
